Qol improvements - Fix test warnings (#111)
* Add shell.nix, update scripts to use /usr/bin/env

* Update docs to work with latest sphinx

* Fix warnings re np.float / np.int

* Remove travis.yml

* Ignore errors from __version__ in setup.py and conf.py

+++ tests/objects/test_matrix33.py
@@ -222,28 +222,28 @@ class test_object_matrix33(unittest.TestCase):
         # add
         self.assertTrue(np.array_equal(m + 1.0, matrix33.create_identity()[:] + 1.0))
         self.assertTrue(np.array_equal(m + 1, matrix33.create_identity()[:] + 1.0))
-        self.assertTrue(np.array_equal(m + np.float(1.), matrix33.create_identity()[:] + 1.0))
+        self.assertTrue(np.array_equal(m + 1., matrix33.create_identity()[:] + 1.0))
         self.assertTrue(np.array_equal(m + fv[0]['f'], matrix33.create_identity()[:] + 2.0))
         self.assertTrue(np.array_equal(m + fv[0]['i'], matrix33.create_identity()[:] + 2.0))
 
         # subtract
         self.assertTrue(np.array_equal(m - 1.0, matrix33.create_identity()[:] - 1.0))
         self.assertTrue(np.array_equal(m - 1, matrix33.create_identity()[:] - 1.0))
-        self.assertTrue(np.array_equal(m - np.float(1.), matrix33.create_identity()[:] - 1.0))
+        self.assertTrue(np.array_equal(m - 1., matrix33.create_identity()[:] - 1.0))
         self.assertTrue(np.array_equal(m - fv[0]['f'], matrix33.create_identity()[:] - 2.0))
         self.assertTrue(np.array_equal(m - fv[0]['i'], matrix33.create_identity()[:] - 2.0))
 
         # multiply
         self.assertTrue(np.array_equal(m * 2.0, matrix33.create_identity()[:] * 2.0))
         self.assertTrue(np.array_equal(m * 2, matrix33.create_identity()[:] * 2.0))
-        self.assertTrue(np.array_equal(m * np.float(2.), matrix33.create_identity()[:] * 2.0))
+        self.assertTrue(np.array_equal(m * 2., matrix33.create_identity()[:] * 2.0))
         self.assertTrue(np.array_equal(m * fv[0]['f'], matrix33.create_identity()[:] * 2.0))
         self.assertTrue(np.array_equal(m * fv[0]['i'], matrix33.create_identity()[:] * 2.0))
 
         # divide
         self.assertTrue(np.array_equal(m / 2.0, matrix33.create_identity()[:] / 2.0))
         self.assertTrue(np.array_equal(m / 2, matrix33.create_identity()[:] / 2.0))
-        self.assertTrue(np.array_equal(m / np.float(2.), matrix33.create_identity()[:] / 2.0))
+        self.assertTrue(np.array_equal(m / 2., matrix33.create_identity()[:] / 2.0))
         self.assertTrue(np.array_equal(m / fv[0]['f'], matrix33.create_identity()[:] / 2.0))
         self.assertTrue(np.array_equal(m / fv[0]['i'], matrix33.create_identity()[:] / 2.0))
