+++ tests/test_ray.py
@@ -15,12 +15,12 @@ class test_ray(unittest.TestCase):
     def test_create(self):
         result = ray.create([0.,0.,0.],[0.,0.,1.])
         np.testing.assert_almost_equal(result, [[0.,0.,0.],[0.,0.,1.]], decimal=5)
-        self.assertTrue(result.dtype == np.float)
+        self.assertTrue(result.dtype == float)
 
     def test_create_dtype(self):
-        result = ray.create([0,0,0],[0,0,1], dtype=np.int)
+        result = ray.create([0,0,0],[0,0,1], dtype=int)
         np.testing.assert_almost_equal(result, [[0,0,0],[0,0,1]], decimal=5)
-        self.assertTrue(result.dtype == np.int)
+        self.assertTrue(result.dtype == int)
 
     def test_create_from_line(self):
         result = ray.create_from_line([
@@ -28,12 +28,12 @@ class test_ray(unittest.TestCase):
             [10.,10.,0.]
         ])
         np.testing.assert_almost_equal(result, [[0.,10.,0.],[1.,0.,0.]], decimal=5)
-        self.assertTrue(result.dtype == np.float)
+        self.assertTrue(result.dtype == float)
 
     def test_invert(self):
         result = ray.invert([[0.,10.,0.],[1.,0.,0.]])
         np.testing.assert_almost_equal(result, [[0.,10.,0.],[-1.,0.,0.]], decimal=5)
-        self.assertTrue(result.dtype == np.float)
+        self.assertTrue(result.dtype == float)
 
     def test_position(self):
         result = ray.position([[0.,10.,0.],[1.,0.,0.]])
